initial commit of a sample Jenkinsfil
this Jenkinsfile is for building, testing and pushing a docker image.
there are built in guards for using samsung cnct production repos,
pushing from master and a way to have a CI job to push explicit
released versions.

please see the kraken-tools release job in Jenkins for how to configure
the jenkins job correclty for releases.